pyproject.toml
[project]
name = "aim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "lpips>=0.1.4",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "transformers>=4.55.0",
    "scikit-learn",
    "accelerate",
    "jupyter>=1.1.1",
    "pandas>=2.3.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
